add speechrecognition

+++ Main.py
@@ -110,37 +110,101 @@ def getFileListFTP(ftp_connection):
 
 def train_and_sync(mode='local'):
     if mode=='local':
-        f_NetworkDump_name = 'NetworkDump_'+retMD5.md5("NetworkDump.json")
-        f_TrainDataDump_name = 'TrainDataDump_'+retMD5.md5("TrainDataDump.pkl")
-        #update current
-        latest_file_list = {'NetworkDump':'','TrainDataDump':'','mnetwork':''}
-        latest_file_list['NetworkDump'] = f_NetworkDump_name
-        latest_file_list['TrainDataDump'] = f_TrainDataDump_name
-        latest_file_list['mnetwork'] = os.path.splitext(retMD5.md5('NetworkDump.json'))[0]+'_'+ os.path.splitext(retMD5.md5('TrainDataDump.pkl'))[0]+".pkl"
-        latest_file_list_file = open('latest_file_list', 'wb')
-        json.dump(latest_file_list,open("latest_file_json",'w'),indent=4)
-
-        #ftp
         try:
+            #ftp stuff
             ftp_connection = ftplib.FTP(server, username, password)
             ftp_connection.cwd(remote_path)
             files = getFileListFTP(ftp_connection)
+
+
+            ##retrieve latest_file_json
+            #from ftp
+            #ftp_connection.storbinary('STOR latest_file_json', open('latest_file_json', 'rb'))
+            try:
+                ftp_connection.retrbinary('RETR %s' % 'latest_file_json', open('ftp_latest_file_json', 'wb').write)
+            except Exception as ex:
+                print(ex)
+            #from localserver
+            shutil.copy2('_History/' + 'latest_file_json', "local_latest_file_json")
+            #from root
+            shutil.copy2('latest_file_json', "root_latest_file_json")
+            #init
+            ftptime=0
+            localtime=0
+            roottime = 0
+            print(json.load(open("latest_file_json", 'r'))['time'])
+            #gettime
+            with open("ftp_latest_file_json", 'r') as f1:
+                ftp_latest_file_json = json.load(f1)
+                try:
+                    ftp_connection.retrbinary('RETR %s' % ftp_latest_file_json['mnetwork'],
+                                              open("NetworkDump_base.pkl", 'wb').write)
+                except Exception as ex:
+                    print(ex)
+            with open("local_latest_file_json", 'r') as f2:
+                local_latest_file_json = json.load(f2)
+            with open("root_latest_file_json", 'r') as f3:
+                root_latest_file_json = json.load(f3)
+            if ftp_latest_file_json['mnetwork'] in files:
+                ftptime = int(ftp_latest_file_json['time'])
+            if local_latest_file_json['mnetwork'] in os.listdir('_History/'):
+                localtime = int(local_latest_file_json['time'])
+            if 'NetworkDump.pkl' in os.listdir('./'):
+                roottime = int(root_latest_file_json['time'])
+            #compare
+            def fetch(mode=''):
+                if mode=='ftp':
+                    fh = open("NetworkDump.pkl", 'wb')
+                    ftp_connection.retrbinary('RETR %s' % ftp_latest_file_json['mnetwork'], fh.write)
+                    print('USE: SERVER')
+                elif mode=='local':
+                    shutil.copy2('_History/' + local_latest_file_json['mnetwork'], "NetworkDump.pkl")
+                    print('USE: LOCAL')
+                elif mode=='root':
+                    print('USE: ROOT')
+                else:
+                    print('USE: UNTOUCHED')
+
+            priortyList = [['ftp',ftptime],['local',localtime],['root',roottime]]
+            priortyList.sort(key=lambda x: x[1],reverse=True)
+
+            def priorController(priorityList):
+                try:
+                    print("trying:"+str(priorityList[0][0]))
+                    fetch(priorityList[0][0])
+                except:
+                    priorController([i for i in priortyList if i[0] != priorityList[0][0]])
+            priorController(priortyList)
+
+
+        except Exception as ex:
+            raise
+            print(ex)
+
+
+        f_NetworkDump_name = 'NetworkDump_' + retMD5.md5("NetworkDump.json")
+        f_TrainDataDump_name = 'TrainDataDump_' + retMD5.md5("TrainDataDump.pkl")
+        # update current
+        latest_file_list = {'time': '', 'NetworkDump': '', 'TrainDataDump': '', 'mnetwork': ''}
+        latest_file_list['time'] = str(int((time.time() + 0.5) * 1000))
+        latest_file_list['NetworkDump'] = f_NetworkDump_name
+        latest_file_list['TrainDataDump'] = f_TrainDataDump_name
+        latest_file_list['mnetwork'] = os.path.splitext(retMD5.md5('NetworkDump.json'))[0] + '_' + \
+                                       os.path.splitext(retMD5.md5('TrainDataDump.pkl'))[0] + ".pkl"
+        ##upload
+        json.dump(latest_file_list, open("latest_file_json", 'w'), indent=4)
+        try:
+            #ftp
             ftp_connection.storbinary('STOR latest_file_json', open('latest_file_json', 'rb'))
             ftp_connection.storbinary('STOR '+f_NetworkDump_name,  open("NetworkDump.json", 'rb'))
             ftp_connection.storbinary('STOR '+f_TrainDataDump_name,  open("TrainDataDump.pkl",'rb'))
-        except:
-            print('FTP error!!')
+        except Exception as ex:
+            print(ex)
         #local
-        shutil.copy2("latest_file_json",'_History/'+"latest_file_json")
-        shutil.copy2("NetworkDump.json",'_History/'+f_NetworkDump_name)
-        shutil.copy2("TrainDataDump.pkl",'_History/'+f_TrainDataDump_name)
+        shutil.copy2("latest_file_json", '_History/' + "latest_file_json")
+        shutil.copy2("NetworkDump.json", '_History/' + f_NetworkDump_name)
+        shutil.copy2("TrainDataDump.pkl", '_History/' + f_TrainDataDump_name)
 
-        #replace if upate exists - ftp
-        if latest_file_list['mnetwork'] in files:
-            fh = open("NetworkDump.pkl", 'rb')
-            ftp_connection.retrbinary('RETR %s' % latest_file_list['mnetwork'], fh.write)
-        #replace if update exists - local
-            shutil.copy2('_History/'+latest_file_list['mnetwork'],"NetworkDump.pkl")
 
     elif mode=='server':
         #FTP stuff
@@ -149,13 +213,20 @@ def train_and_sync(mode='local'):
         files = getFileListFTP(ftp_connection)
 
         ftp_connection.retrbinary('RETR %s' % 'latest_file_json', open('latest_file_json', 'wb').write)
+        latest_file_list = json.load(open("latest_file_json", 'r'))
+        print(latest_file_list)
+        ftp_connection.retrbinary('RETR %s' % latest_file_list['NetworkDump'],  open("NetworkDump.json", 'wb').write)
+        ftp_connection.retrbinary('RETR %s' % latest_file_list['TrainDataDump'], open('TrainDataDump.pkl', 'wb').write)
         try:
             latest_file_list = json.load(open("latest_file_json",'r'))
             print(latest_file_list['mnetwork'])
             if latest_file_list['mnetwork'] not in files:
                 RNNinterface([],[], json2network(), Mode='Train',
                              PLKnetwork_PATH='NetworkDump.pkl', train=True,maxEpochs=None)
-                ftp_connection.storbinary('STOR '+str(latest_file_list['mnetwork']),open(Paths.get('pklNetworkDumo'),'rb'))
+                ftp_connection.storbinary('STOR ' + str(latest_file_list['mnetwork']),
+                                          open(Paths.get('pklNetworkDumo'), 'rb'))
+                ftp_connection.storbinary('STOR ' + str('latest_file_json'),
+                                          open('latest_file_json', 'rb'))
         except Exception as ex:
             print(ex)
     elif mode == 'localserver':
@@ -165,8 +236,9 @@ def train_and_sync(mode='local'):
 
             if latest_file_list['mnetwork'] not in os.listdir('_History/'):
                 RNNinterface([],[], json2network(), Mode='Train',
-                             PLKnetwork_PATH='NetworkDump.pkl', train=True,maxEpochs=None)
+                             PLKnetwork_PATH='NetworkDump.pkl', train=True,maxEpochs=20)
                 shutil.copy2("NetworkDump.pkl",'_History/'+latest_file_list['mnetwork'])
+                shutil.copy2("latest_file_json", '_History/' + 'latest_file_json')
         except Exception as ex:
             print(ex)
 
@@ -187,29 +259,33 @@ if __name__ == "__main__":
                 try:
                     curr = json.load(open("latest_file_json",'r'))['mnetwork']
                 except:
+                    time.sleep(5)
                     print('latest_file_json ERROR')
-                    continue
-
+                    pass
                 while p.is_alive():
-                    time.sleep(1)
-                    print('testing...')
-
+                    time.sleep(10)
+                    #print('testing...')
                     filelist =  os.listdir('_History/')
                     try:
                         latest_file_list = json.load(open('latest_file_json','r'))
-                        print(curr)
-                        print(latest_file_list['mnetwork'])
+                        #print(curr)
+                        #print(latest_file_list['mnetwork'])
                         if curr!=latest_file_list['mnetwork']:
                             print("THERE IS NEW")
-                            p.terminate()
+                            if train_and_sync_mode == 'localserver':
+                                p.terminate()
+                                print('INTERRUPTED')
                     except:
+                        time.sleep(10)
                         continue
             except Exception as ex:
+                time.sleep(10)
                 print(ex)
                 continue
 
     while train_and_sync_mode == 'local':
-        print('local')
+        mtrainandsync_local = trainandsync_local()
+        mtrainandsync_local.start()
         newInputs = []
         mnewinput = None
 
@@ -239,8 +315,18 @@ if __name__ == "__main__":
 
                 # train others
                 RNNinterface(newInput.subjectList, newInput.getreactionList(), json2network(), Mode='Activate')
-                RNNinterface(newInput.getsubjectList(), newInput.getreactionList(), json2network(), Mode='Train',
-                             PLKnetwork_PATH=Paths.get('pklNetworkDumo'), train=newInput.getTrainConfig())
+                if newInput.getTrainConfig() == True:
+                    RNNinterface(newInput.getsubjectList(), newInput.getreactionList(), json2network(), Mode='Train',
+                             PLKnetwork_PATH=Paths.get('pklNetworkDumo'), train=True,maxEpochs=200)
+                    latest_file_list = json.load(open("latest_file_json", 'r'))
+                    latest_file_list['time'] = str(int((time.time() + 0.5) * 1000))
+                    with open("latest_file_json", 'w') as outfile:
+                        json.dump(latest_file_list, outfile, indent=4)
+                    print('end')
+                    print(json.load(open("latest_file_json",'r'))['time'])
+                elif newInput.getTrainConfig() == False:
+                    RNNinterface(newInput.getsubjectList(), newInput.getreactionList(), json2network(), Mode='Train',
+                                 PLKnetwork_PATH=Paths.get('pklNetworkDumo'), train=False)
 
         # if Activation Mode
         else:
@@ -257,7 +343,6 @@ if __name__ == "__main__":
 
             # Activate
 
-        mtrainandsync_local = trainandsync_local()
-        mtrainandsync_local.start()
+
